Simulation

+++ Microsimulation/Entities/BirthProbability.cs
@@ -9,7 +9,7 @@ namespace Microsimulation.Entities
     public class BirthProbability
     {
         public int Age { get; set; }
-        public byte NbrOfChildren { get; set; }
+        public int NbrOfChildren { get; set; }
         public double Probability { get; set; }
     }
 }
